+++ PortalECI/src/it/portalECI/bo/GestioneTipiVerificaBO.java
@@ -1,5 +1,6 @@
 package it.portalECI.bo;
 
+import org.hibernate.Query;
 import org.hibernate.Session;
 
 import it.portalECI.DAO.GestioneTipiVerificaDAO;
@@ -39,13 +40,19 @@ public class GestioneTipiVerificaBO {
 	}
 	
 	public static int deleteTipoVerifica(TipoVerificaDTO tipo, Session session) {
+		Query query = session.createQuery("from VerbaleDTO WHERE codiceVerifica = :_codiceVerifica");
+		query.setParameter("_codiceVerifica",tipo.getCodice());
+		
+		if(query.list().size()>0){	
+			return 2;
+		}
+		
 		int toRet=0;
 		try {
 			session.delete(tipo);
 			toRet=0;
 		}catch (Exception ex){
 			toRet=1;
-			throw ex;	 			 		
 		}
 		return toRet;	
 	}
